Refactor auth routes into per-endpoint modules
